create static resource table lib
