+++ lib/resource/table.c
@@ -1,0 +1,67 @@
+// Copyright Joshua MacDonald
+// SPDX-License-Identifier: MIT
+
+// TODO: This is a placeholder.  Presently there is nothing
+// configurable here--all the examples use the same resource table.
+
+#include <stddef.h>
+
+#include "external/ti-pru-support/include/pru_virtio_ids.h"
+#include "external/ti-pru-support/include/rsc_types.h"
+
+#include "table.h"
+
+#define NUM_RESOURCES 1
+
+// Sizes of the virtqueues (expressed in number of buffers supported,
+// and must be power of 2)
+#define PRU_RPMSG_VQ0_SIZE 16
+#define PRU_RPMSG_VQ1_SIZE 16
+
+// The feature bitmap for virtio rpmsg
+#define VIRTIO_RPMSG_F_NS 0 // name service notifications
+
+// This firmware supports name service notifications as one of its features.
+#define RPMSG_PRU_C0_FEATURES (1 << VIRTIO_RPMSG_F_NS)
+
+#pragma DATA_SECTION(resourceTable, ".resource_table")
+#pragma RETAIN(resourceTable)
+struct my_resource_table resourceTable = {
+    // resource_table base
+    {
+        1,             // Resource table version: only version 1 is supported
+        NUM_RESOURCES, // Number of entries in the table (equals length of offset field).
+        0, 0,          // Reserved zero fields
+    },
+    // Entry offsets
+    {
+        offsetof(struct my_resource_table, rpmsg_vdev),
+    },
+    // RPMsg virtual device
+    {
+        (uint32_t)TYPE_VDEV,             // type
+        (uint32_t)VIRTIO_ID_RPMSG,       // id
+        (uint32_t)0,                     // notifyid
+        (uint32_t)RPMSG_PRU_C0_FEATURES, // dfeatures
+        (uint32_t)0,                     // gfeatures
+        (uint32_t)0,                     // config_len
+        (uint8_t)0,                      // status
+        (uint8_t)2,                      // num_of_vrings, only two is supported
+        {(uint8_t)0, (uint8_t)0},        // reserved
+    },
+    // The two vring structs must be packed after the vdev entry.
+    {
+        FW_RSC_ADDR_ANY,    // da, will be populated by host, can't pass it in
+        16,                 // align (bytes),
+        PRU_RPMSG_VQ0_SIZE, // num of descriptors
+        0,                  // notifyid, will be populated, can't pass right now
+        0                   // reserved
+    },
+    {
+        FW_RSC_ADDR_ANY,    // da, will be populated by host, can't pass it in
+        16,                 // align (bytes),
+        PRU_RPMSG_VQ1_SIZE, // num of descriptors
+        0,                  // notifyid, will be populated, can't pass right now
+        0                   // reserved
+    },
+};
